volume should be long, not int

+++ src/StockScreenerReports.Storage/Reports.fs
@@ -32,7 +32,7 @@ module Reports =
             marketCap:decimal;
             price:decimal;
             change:decimal;
-            volume:int;
+            volume:int64;
             screenerid:int;
             screenername:string;
         }
